requirements.txt
langchain
langchain_community
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
python-dotenv


sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
langchain-openai

streamlit
groq
langchain_groq
langchain_core
fastapi
uvicorn
langserve[all]